Methoden/Klassen für Index-Nattable auslagern
